Fixes to allow mpe:: calls even when MPE is not enabled
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@4113 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/adlb/tcl-adlb.c
@@ -338,7 +338,8 @@ ADLB_Get_Cmd(ClientData cdata, Tcl_Interp *interp,
 /**
    Convert type string to adlb_data_type
  */
-static inline adlb_data_type type_from_string(char* type_string)
+static inline
+adlb_data_type type_from_string(char* type_string)
 {
   adlb_data_type result;
   if (strcmp(type_string, "integer") == 0)
@@ -502,7 +503,7 @@ ADLB_Store_Cmd(ClientData cdata, Tcl_Interp *interp,
         int p;
         rc = Tcl_GetIntFromObj(interp, objv[3], &p);
         TCL_CHECK_MSG(rc, "required pointer!");
-        data = (void*) p;
+        data = (void*) (long) p;
         rc = Tcl_GetIntFromObj(interp, objv[4], &length);
         TCL_CHECK_MSG(rc, "required length!");
       }
